Remove fan_common module
Inlines the necessary enums inside integrations they were used.
